# src/main/java/io/dudsthekid3756/schema_validator/domain/application/CreateResponse.java
package io.dudsthekid3756.schema_validator.domain.application;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

@Getter
@Setter
@ToString
@JsonIgnoreProperties(ignoreUnknown = true)
public class CreateResponse {
    private String id;
    private String status;
}